Parse and show more information on posts, clean up the layout a bitsy

+++ Revibe/src/com/revibe/utils/DateHelper.java
@@ -32,7 +32,7 @@ public class DateHelper {
         if (seconds == 1)
             return "a second ago";
 
-        return seconds + " seconds ago";
+        return (int)seconds + " seconds ago";
     }
 
 }
